feat(posts): restrict post edits and deletions to post owners

+++ post/views.py
@@ -7,9 +7,8 @@ from .models import Post
 
 @login_required(login_url="/users/signin")
 def posts(request):
-    print("User:", request)
-    posts = Post.objects.filter(author=request.user)
-    # print(posts)
+    # posts = Post.objects.filter(author=request.user)
+    posts = Post.objects.all()
     return render(request, "post/dashboard.html", {"posts": posts})
 
 
@@ -33,18 +32,26 @@ def add_post(request):
 
 @login_required(login_url="/users/signin")
 def edit_post(request, id):
+    post = Post.objects.get(id=id)
+
+    if post.author != request.user:
+        return redirect("post:posts")
+
     if request.method == "POST":
         title = request.POST["title"]
         content = request.POST["content"]
         Post.objects.filter(id=id).update(title=title, content=content)
         return redirect("post:post_detail", id=id)
 
-    post = Post.objects.get(id=id)
     return render(request, "post/edit_post.html", {"post": post})
 
 
 @login_required(login_url="/users/signin")
 def delete_post(request, id):
     post = Post.objects.get(id=id)
+
+    if post.author != request.user:
+        return redirect("post:posts")
+
     post.delete()
     return redirect("post:posts")
